+++ src/components/index.ts
@@ -1,2 +1,5 @@
 export * from './Title';
 export * from './Input';
+export * from './Header';
+export * from './SideNavigation';
+export * from './Layouts';
